feat: Account Deletion pop up (READ DESCRIPTION ON HOW TO USE IT RN)
Right now the pop up is assigned to the get Started button, and look weird, but that it is because the welcome screen is not structured for the button. (we center inside the view that is in the second half of the screen). So it will be properley centered, or can be adjusted when we have the settings page. If merging! We should unassign from get started button (just uncomment)